+++ client/packages/cli/index.js
@@ -1,7 +1,6 @@
 // @ts-check
 
 import version from "./src/version.js";
-import setupHelp from "./src/setupHelp.js";
 import { mkdir, writeFile, readFile, stat } from "fs/promises";
 import { join } from "path";
 import { randomUUID } from "crypto";
@@ -41,8 +40,26 @@ const instantBackendOrigin =
 
 // cli
 
-setupHelp(program);
 
+
+// Header -- this shows up in every command
+const logoChalk = chalk.bold("instant-cli");
+const versionChalk = chalk.dim(`${version.trim()}`);
+const headerChalk = `${logoChalk} ${versionChalk} ` + "\n";
+
+
+// Help Footer -- this only shows up in help commands
+const helpFooterChalk =
+  "\n" +
+  `
+${chalk.bold.dim("Want to learn more?")}
+${chalk.white("Check out the docs")}: ${chalk.blueBright.underline("https://instantdb.com/docs")}
+${chalk.white("Join the Discord")}:   ${chalk.blueBright.underline("https://discord.com/invite/VU53p7uQcE")}
+`.trim();
+
+
+program.addHelpText("after", helpFooterChalk);
+program.addHelpText("beforeAll", headerChalk);
 program
   .name("instant-cli")
   .option("-t --token <TOKEN>", "auth token override")
